test(desktop): consolidate agent studio and orchestrator test fixtures
