+++ apps/desktop/src/pages/use-agent-studio-page-models.test.tsx
@@ -1,67 +1,34 @@
 import { describe, expect, mock, test } from "bun:test";
 import type { TaskDocumentState } from "@/components/features/task-details/use-task-documents";
-import type { AgentSessionState } from "@/types/agent-orchestrator";
-import type { TaskCard } from "@openducktor/contracts";
-import { type ReactElement, createElement } from "react";
-import TestRenderer, { act } from "react-test-renderer";
+import { act } from "react-test-renderer";
+import {
+  createAgentSessionFixture,
+  createHookHarness as createSharedHookHarness,
+  createTaskCardFixture,
+  enableReactActEnvironment,
+} from "./agent-studio-test-utils";
 import { useAgentStudioPageModels } from "./use-agent-studio-page-models";
 
-const reactActEnvironment = globalThis as typeof globalThis & {
-  IS_REACT_ACT_ENVIRONMENT?: boolean;
-};
-reactActEnvironment.IS_REACT_ACT_ENVIRONMENT = true;
+enableReactActEnvironment();
 
 type HookArgs = Parameters<typeof useAgentStudioPageModels>[0];
 type HookState = ReturnType<typeof useAgentStudioPageModels>;
 
-const createTask = (): TaskCard => ({
-  id: "task-1",
-  title: "Task 1",
-  description: "",
-  acceptanceCriteria: "",
-  notes: "",
-  status: "open",
-  priority: 2,
-  issueType: "task",
-  aiReviewEnabled: true,
-  availableActions: [],
-  labels: [],
-  parentId: undefined,
-  subtaskIds: [],
-  assignee: undefined,
-  documentSummary: {
-    spec: { has: true },
-    plan: { has: false },
-    qaReport: { has: false },
-  },
-  updatedAt: "2026-02-22T12:00:00.000Z",
-  createdAt: "2026-02-22T12:00:00.000Z",
-});
+const createTask = () =>
+  createTaskCardFixture({
+    documentSummary: {
+      spec: { has: true },
+      plan: { has: false },
+      qaReport: { has: false },
+    },
+  });
 
-const createSession = (
-  sessionId = "session-1",
-  externalSessionId = "external-1",
-): AgentSessionState => ({
-  sessionId,
-  externalSessionId,
-  taskId: "task-1",
-  role: "spec",
-  scenario: "spec_initial",
-  status: "running",
-  startedAt: "2026-02-22T10:00:00.000Z",
-  runtimeId: null,
-  runId: null,
-  baseUrl: "http://localhost:4000",
-  workingDirectory: "/repo",
-  messages: [],
-  draftAssistantText: "",
-  pendingPermissions: [],
-  pendingQuestions: [],
-  todos: [],
-  modelCatalog: null,
-  selectedModel: null,
-  isLoadingModelCatalog: false,
-});
+const createSession = (sessionId = "session-1", externalSessionId = "external-1") =>
+  createAgentSessionFixture({
+    sessionId,
+    externalSessionId,
+    status: "running",
+  });
 
 const createDocumentState = (markdown: string): TaskDocumentState => ({
   markdown,
@@ -71,53 +38,8 @@ const createDocumentState = (markdown: string): TaskDocumentState => ({
   loaded: true,
 });
 
-const flush = async (): Promise<void> => {
-  await Promise.resolve();
-  await Promise.resolve();
-};
-
-const createHookHarness = (initialProps: HookArgs) => {
-  let latest: HookState | null = null;
-  let currentProps = initialProps;
-
-  const Harness = (props: HookArgs): ReactElement | null => {
-    latest = useAgentStudioPageModels(props);
-    return null;
-  };
-
-  let renderer: TestRenderer.ReactTestRenderer | null = null;
-
-  const mount = async (): Promise<void> => {
-    await act(async () => {
-      renderer = TestRenderer.create(createElement(Harness, currentProps));
-      await flush();
-    });
-  };
-
-  const update = async (nextProps: HookArgs): Promise<void> => {
-    currentProps = nextProps;
-    await act(async () => {
-      renderer?.update(createElement(Harness, currentProps));
-      await flush();
-    });
-  };
-
-  const getLatest = (): HookState => {
-    if (!latest) {
-      throw new Error("Hook state unavailable");
-    }
-    return latest;
-  };
-
-  const unmount = async (): Promise<void> => {
-    await act(async () => {
-      renderer?.unmount();
-      await flush();
-    });
-  };
-
-  return { mount, update, getLatest, unmount };
-};
+const createHookHarness = (initialProps: HookArgs) =>
+  createSharedHookHarness(useAgentStudioPageModels, initialProps);
 
 describe("useAgentStudioPageModels", () => {
   test("builds page models and forwards wrapper callbacks", async () => {
@@ -264,7 +186,6 @@ describe("useAgentStudioPageModels", () => {
 
     await act(async () => {
       harness.getLatest().agentChatModel.thread.onToggleTodoPanel();
-      await flush();
     });
     expect(harness.getLatest().agentChatModel.thread.todoPanelCollapsed).toBe(true);
 
@@ -277,7 +198,6 @@ describe("useAgentStudioPageModels", () => {
 
     await act(async () => {
       harness.getLatest().agentChatModel.thread.onToggleTodoPanel();
-      await flush();
     });
     expect(harness.getLatest().agentChatModel.thread.todoPanelCollapsed).toBe(true);
