+++ apps/desktop/src/pages/use-agent-studio-task-tabs.test.tsx
@@ -1,16 +1,15 @@
 import { describe, expect, mock, test } from "bun:test";
-import type { AgentSessionState } from "@/types/agent-orchestrator";
-import type { TaskCard } from "@openducktor/contracts";
-import { type ReactElement, createElement } from "react";
-import TestRenderer, { act } from "react-test-renderer";
+import {
+  createAgentSessionFixture,
+  createHookHarness as createSharedHookHarness,
+  createTaskCardFixture,
+  enableReactActEnvironment,
+} from "./agent-studio-test-utils";
 import { toPersistedTaskTabs } from "./agents-page-session-tabs";
 import { toTabsStorageKey } from "./agents-page-utils";
 import { useAgentStudioTaskTabs } from "./use-agent-studio-task-tabs";
 
-const reactActEnvironment = globalThis as typeof globalThis & {
-  IS_REACT_ACT_ENVIRONMENT?: boolean;
-};
-reactActEnvironment.IS_REACT_ACT_ENVIRONMENT = true;
+enableReactActEnvironment();
 
 type HookArgs = Parameters<typeof useAgentStudioTaskTabs>[0];
 type HookState = ReturnType<typeof useAgentStudioTaskTabs>;
@@ -39,101 +38,17 @@ const createMemoryStorage = (): StorageLike => {
   };
 };
 
-const createTask = (id: string): TaskCard => ({
-  id,
-  title: id,
-  description: "",
-  acceptanceCriteria: "",
-  notes: "",
-  status: "open",
-  priority: 2,
-  issueType: "task",
-  aiReviewEnabled: true,
-  availableActions: [],
-  labels: [],
-  parentId: undefined,
-  subtaskIds: [],
-  assignee: undefined,
-  documentSummary: {
-    spec: { has: false },
-    plan: { has: false },
-    qaReport: { has: false },
-  },
-  updatedAt: "2026-02-22T12:00:00.000Z",
-  createdAt: "2026-02-22T12:00:00.000Z",
-});
+const createTask = (id: string) => createTaskCardFixture({ id, title: id });
 
-const createSession = (taskId: string, sessionId: string): AgentSessionState => ({
-  sessionId,
-  externalSessionId: `ext-${sessionId}`,
-  taskId,
-  role: "spec",
-  scenario: "spec_initial",
-  status: "idle",
-  startedAt: "2026-02-22T10:00:00.000Z",
-  runtimeId: null,
-  runId: null,
-  baseUrl: "http://localhost:4000",
-  workingDirectory: "/repo",
-  messages: [],
-  draftAssistantText: "",
-  pendingPermissions: [],
-  pendingQuestions: [],
-  todos: [],
-  modelCatalog: null,
-  selectedModel: null,
-  isLoadingModelCatalog: false,
-});
+const createSession = (taskId: string, sessionId: string) =>
+  createAgentSessionFixture({
+    sessionId,
+    externalSessionId: `ext-${sessionId}`,
+    taskId,
+  });
 
-const flush = async (): Promise<void> => {
-  await Promise.resolve();
-  await Promise.resolve();
-};
-
-const createHookHarness = (initialProps: HookArgs) => {
-  let latest: HookState | null = null;
-  const currentProps = initialProps;
-
-  const Harness = (props: HookArgs): ReactElement | null => {
-    latest = useAgentStudioTaskTabs(props);
-    return null;
-  };
-
-  let renderer: TestRenderer.ReactTestRenderer | null = null;
-
-  const mount = async (): Promise<void> => {
-    await act(async () => {
-      renderer = TestRenderer.create(createElement(Harness, currentProps));
-      await flush();
-    });
-  };
-
-  const run = async (fn: (state: HookState) => void | Promise<void>): Promise<void> => {
-    await act(async () => {
-      if (!latest) {
-        throw new Error("Hook state unavailable");
-      }
-      await fn(latest);
-      await flush();
-    });
-  };
-
-  const getLatest = (): HookState => {
-    if (!latest) {
-      throw new Error("Hook state unavailable");
-    }
-    return latest;
-  };
-
-  const unmount = async (): Promise<void> => {
-    await act(async () => {
-      renderer?.unmount();
-      await flush();
-    });
-  };
-
-  return { mount, run, getLatest, unmount };
-};
+const createHookHarness = (initialProps: HookArgs) =>
+  createSharedHookHarness(useAgentStudioTaskTabs, initialProps);
 
 describe("useAgentStudioTaskTabs", () => {
   test("hydrates persisted task tabs from localStorage", async () => {
