+++ apps/desktop/src/state/operations/agent-orchestrator/test-utils.ts
@@ -1,0 +1,79 @@
+import type { AgentSessionState } from "@/types/agent-orchestrator";
+import type { TaskCard } from "@openducktor/contracts";
+
+export const createDeferred = <T>() => {
+  let resolve: ((value: T | PromiseLike<T>) => void) | null = null;
+  let reject: ((reason?: unknown) => void) | null = null;
+  const promise = new Promise<T>((res, rej) => {
+    resolve = res;
+    reject = rej;
+  });
+
+  return {
+    promise,
+    resolve: (value: T) => resolve?.(value),
+    reject: (reason?: unknown) => reject?.(reason),
+  };
+};
+
+export const withTimeout = async <T>(
+  promise: Promise<T>,
+  timeoutMs: number,
+): Promise<T | "timeout"> => {
+  return Promise.race([
+    promise,
+    new Promise<"timeout">((resolve) => {
+      setTimeout(() => resolve("timeout"), timeoutMs);
+    }),
+  ]);
+};
+
+export const createTaskCardFixture = (overrides: Partial<TaskCard> = {}): TaskCard => ({
+  id: "task-1",
+  title: "Task",
+  description: "",
+  acceptanceCriteria: "",
+  notes: "",
+  status: "open",
+  priority: 2,
+  issueType: "task",
+  aiReviewEnabled: true,
+  availableActions: [],
+  labels: [],
+  parentId: undefined,
+  subtaskIds: [],
+  assignee: undefined,
+  documentSummary: {
+    spec: { has: false },
+    plan: { has: false },
+    qaReport: { has: false },
+  },
+  updatedAt: "2026-02-22T08:00:00.000Z",
+  createdAt: "2026-02-22T08:00:00.000Z",
+  ...overrides,
+});
+
+export const createAgentSessionFixture = (
+  overrides: Partial<AgentSessionState> = {},
+): AgentSessionState => ({
+  sessionId: "session-1",
+  externalSessionId: "external-1",
+  taskId: "task-1",
+  role: "spec",
+  scenario: "spec_initial",
+  status: "idle",
+  startedAt: "2026-02-22T08:00:00.000Z",
+  runtimeId: null,
+  runId: null,
+  baseUrl: "http://127.0.0.1:4444",
+  workingDirectory: "/tmp/repo/worktree",
+  messages: [],
+  draftAssistantText: "",
+  pendingPermissions: [],
+  pendingQuestions: [],
+  todos: [],
+  modelCatalog: null,
+  selectedModel: null,
+  isLoadingModelCatalog: false,
+  ...overrides,
+});
